fix: regression introduce in previous release

+++ src/Config.php
@@ -458,7 +458,7 @@ class Config
     {
         if ($themes = $this->getTheme()) {
             foreach ($themes as $theme) {
-                if (!Util\File::getFS()->exists($this->getThemeDirPath($theme, 'layouts')) || !Util\File::getFS()->exists(Util::joinFile($this->getThemesPath(), $theme, 'config.yml'))) {
+                if (!Util\File::getFS()->exists($this->getThemeDirPath($theme, 'layouts')) && !Util\File::getFS()->exists(Util::joinFile($this->getThemesPath(), $theme, 'config.yml'))) {
                     throw new RuntimeException(\sprintf('Theme "%s" not found. Did you forgot to install it?', $theme));
                 }
             }
